refactor(InputView) : InputView 수합
